server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
scikit-learn==1.5.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0